fix: 자료형을 int에서 long으로 변경

+++ src/main/java/board/BoardDao.java
@@ -38,7 +38,7 @@ public class BoardDao {
         boardDto.setId(keyValue.longValue());
     }
 
-    public List<BoardDto> searchById(int index) {
+    public List<BoardDto> searchById(long index) {
         List<BoardDto> result = jdbcTemplate.query("select * from board where id=?",
                 (ResultSet rs, int numRow) -> resultBoardDto(rs), index);
         return result.isEmpty() ? null : result;
